Prevents race condition when identifying a Task pool in `update`
When first attempting a task in `update` (but not in `init`) the sub tasks were not run. The reason for this is because the model was not updated in time for the identify message.

Adding Promise.resolve around the send command allows a tick to happen so elm can update the model in time